feat: enhance wind farm controls with disassembly features and UI updates
- Updated index.html to include new buttons for disassembly and assembly of turbine components.
- Modified main.mjs to handle new disassembly functionality, including focus adjustments and state management.
- Enhanced turbine selection logic to reflect disassembly state and added input for disassembly degree.
- Improved validation tests in test_controls.mjs to cover new disassembly features and ensure correct behavior.
- Updated wind_farm_controls.mjs to manage disassembly states and apply visual changes during inspections.
- Added new validation script for disassembly checks and generated reports for validation results.
- Updated requirements.txt to include threadpoolctl for improved performance.
- Adjusted API endpoints in main.py to use Query parameters for hours with validation.
- Enhanced weather provider to ensure proper DataFrame creation with float dtype.
- Added new weather data JSON file for testing and validation purposes.

+++ src/services/weather_provider.py
@@ -99,7 +99,7 @@ class OpenMeteoProvider:
             hourly = package['response']['hourly']
             frame = pd.DataFrame({'wind_speed': hourly[WIND_VARIABLE+suffix],
                                   'temperature': hourly['temperature_2m'+suffix]},
-                                 index=pd.to_datetime(hourly['time'], utc=True))
+                                 index=pd.to_datetime(hourly['time'], utc=True), dtype=float)
             frame.index.name = 'timestamp'
             return frame
         except (KeyError, ValueError) as exc:
